Easy Logic
Make two arrays, left and right. In the left array, store the product values that are to the left of the particular element in the main array. Same logic for right array, but storing product values that are to the right of each element of the main array. Finally, iterate through the main array and find the product of every element of the array except itself by multiplying left[i] * right[i]