USING BFS
This uses another type of iteration of adjacent nodes of a particular. Here, each node is first popped from the list, and its adjacent nodes are traversed, and the distances in each of them are modified accordingly. Finally, the valid adjacent nodes are appended to the queue. Such, a traversal can also be used for directed graphs also. (not yet tried with bst)