USING DFS (PLS REVISIT)
Here, since there can be multiple paths between the start and the end node, the visited matrix keeps track of the nodes that have been visited for a particular path. So, suppose all the four directions from a particular node is traversed, then that node is set to unvisited again and the recursion backtracks to the previous call. (I have written down the rough work for the traversal of the graph)