USING RECURSION (PLS REVISIT)
In each recursive call, the string is operated in two ways. First is, either the first character of the string is added to the output, and another recursive call is made with this new output, along with passing the string without the first character. Thus, this process is repeated until the length of the string becomes zero. That is when the output is added to the set.