USING SORTING AND TWO POINTER APPROACH
In two sum, we sort the array and then use two pointer approach to check if the sum of the two elements is greater, smaller or equal to the required sum. If equal, then return True. Otherwise, if lesser than that, then taking the pointer at/near the starting, we increment it to a unique element of the array. If the sum is greater than the req. sum, then taking pointer at/near the end, we decrement it to a unique element of the array. Now, the same operation is used in this problem, only difference is that another outer for loop is used to keep track of the 3rd element, which is part of the triplet.